Changes to the architecture
This commit changes many things in the architecture so that other legacy networks could be added easyly and without duplicating too much server code, for that has been added the following interfaces and classes:
	* LegacyNetwork - this abstract class defines the methods that every legacy network MUST implement, this way we generalize the common methods used by the server.
	* LegacyContact - A place holder of the contact information, we this class we generalize the way the server gets the user information, every legacy network should convert their contact information to this format.
	* LegacyUserStatus - Defines the value of user's status.

The following are the callbacks interfaces. This interfaces are implemented by the server and any legacy network MUST inform the server of any events through the callbacks.
	* ContactListCallbacks - Callbacks for contact list events.
	* LoginCallbacks - Callbacks for login events.
	* MessageCallbacks - Callbacks for message events.

This commits deprecates and removes the MPIMMessenger class, we no longer use this class in favor to the LegacyMsn that follows the new convention, although not all the functionality is added the server is unable to retrieve a contact avatar.