Some fixes and use of another Escaping method
The bug with the update of the presence has been corrected, and now we no longer create the stanzas manually in favor to the methods of PresenceStanza class. I have also corrected the problem when a client sends a html message to a legacy network, for some unknown reason the Xml parser could not parse the encoded string so I have added a work around, is not pretty but gets the job done. I have decided to no longer use the Util.encodeHTML method in favor to the StringEscapeUtils from the apache commons lang package, this means that the class Util no longer is in the directory structure.